+++ src/controllers/setup_stats/down_stats_game.c
@@ -11,7 +11,7 @@ extern int class_base_data[3][5];
 
 void down_strength_game(game_t *game)
 {
-    if (game->saves[game->current]->stats->strength - 1 <
+    if ((int) game->saves[game->current]->stats->strength - 1 <
         class_base_data[game->saves[game->current]->class][0])
         return;
     game->saves[game->current]->stats->strength--;
@@ -24,7 +24,7 @@ void down_strength_game(game_t *game)
 
 void down_defense_game(game_t *game)
 {
-    if (game->saves[game->current]->stats->defense - 1 <
+    if ((int) game->saves[game->current]->stats->defense - 1 <
         class_base_data[game->saves[game->current]->class][1])
         return;
     game->saves[game->current]->stats->defense--;
@@ -37,7 +37,7 @@ void down_defense_game(game_t *game)
 
 void down_vitality_game(game_t *game)
 {
-    if (game->saves[game->current]->stats->vitality - 1 <
+    if ((int) game->saves[game->current]->stats->vitality - 1 <
         class_base_data[game->saves[game->current]->class][2])
         return;
     game->saves[game->current]->stats->vitality--;
@@ -50,7 +50,7 @@ void down_vitality_game(game_t *game)
 
 void down_speed_game(game_t *game)
 {
-    if (game->saves[game->current]->stats->speed - 1 <
+    if ((int) game->saves[game->current]->stats->speed - 1 <
         class_base_data[game->saves[game->current]->class][3])
         return;
     game->saves[game->current]->stats->speed--;
@@ -63,7 +63,7 @@ void down_speed_game(game_t *game)
 
 void down_dexterity_game(game_t *game)
 {
-    if (game->saves[game->current]->stats->dexterity - 1 <
+    if ((int) game->saves[game->current]->stats->dexterity - 1 <
         class_base_data[game->saves[game->current]->class][4])
         return;
     game->saves[game->current]->stats->dexterity--;
